+++ src/victory-util/continuous-transitions.js
@@ -1,60 +1,12 @@
 /* eslint-disable func-style */
 import { assign, defaults, identity, filter } from "lodash";
 import React from "react";
-
-function getDatumKey(datum, idx) {
-  return (datum.key || idx).toString();
-}
-
-function getKeyedData(data) {
-  return data.reduce((keyedData, datum, idx) => {
-    const key = getDatumKey(datum, idx);
-    keyedData[key] = datum;
-    return keyedData;
-  }, {});
-}
-
-function getKeyedDataDifference(a, b) {
-  let hasDifference = false;
-  const difference = Object.keys(a).reduce((_difference, key) => {
-    if (!(key in b)) {
-      hasDifference = true;
-      _difference[key] = true;
-    }
-    return _difference;
-  }, {});
-  return hasDifference && difference;
-}
-
-/**
- * Calculate which data-points exist in oldData and not nextData -
- * these are the `entering` data-points.  Also calculate which
- * data-points exist in nextData and not oldData - these are the
- * `exiting` data-points.
- *
- * @param  {Array} oldData   this.props.data Array
- * @param  {Array} nextData  this.props.data Array
- *
- * @return {Object}          Object with `exiting` and `entering` properties.
- *                           exiting[datum.key] will be true if the data is
- *                           exiting, and similarly for `entering`.
- */
-function getNodeTransitions(oldData, nextData) {
-  const oldDataKeyed = oldData && getKeyedData(oldData);
-  const nextDataKeyed = nextData && getKeyedData(nextData);
-
-  return {
-    entering: oldDataKeyed && getKeyedDataDifference(nextDataKeyed, oldDataKeyed),
-    exiting: nextDataKeyed && getKeyedDataDifference(oldDataKeyed, nextDataKeyed)
-  };
-}
-
-function getChildData(child) {
-  if (child.type && child.type.getData) {
-    return child.type.getData(child.props);
-  }
-  return child.props && child.props.data || false;
-}
+import {
+  getDatumKey,
+  getNodeTransitions,
+  getChildData,
+  getChildTransitionDuration
+} from "./transition-helpers";
 
 /**
  * If a parent component has animation enabled, calculate the transitions
@@ -72,9 +24,10 @@ function getChildData(child) {
  *                                    - childrenTransitions
  *                                    - nodesShouldExit
  */
+
 export function getInitialTransitionState(oldChildren, nextChildren) {
-  let nodesWillEnter = false;
   let nodesWillExit = false;
+  let nodesWillEnter = false;
 
   const getTransition = (oldChild, newChild) => {
     if (!newChild || oldChild.type !== newChild.type) {
@@ -229,13 +182,6 @@ export function getTransitionPropsFactory(props, state, setState) {
       getChildPropsBeforeExit(animate, data, nodes, () => {
         setState({ nodesShouldExit: true });
       });
-  };
-
-  const getChildTransitionDuration = function (child, type) {
-    const animate = child.props.animate;
-    const defaultTransitions = child.type && child.type.defaultTransitions;
-    return animate[type] && animate[type].duration ||
-      defaultTransitions[type] && defaultTransitions[type].duration;
   };
 
   return function getTransitionProps(child, index) { // eslint-disable-line max-statements
