+++ src/victory-util/transitions.js
@@ -1,60 +1,12 @@
 /* eslint-disable func-style */
 import { assign, defaults, identity } from "lodash";
 import React from "react";
-
-function getDatumKey(datum, idx) {
-  return (datum.key || idx).toString();
-}
-
-function getKeyedData(data) {
-  return data.reduce((keyedData, datum, idx) => {
-    const key = getDatumKey(datum, idx);
-    keyedData[key] = datum;
-    return keyedData;
-  }, {});
-}
-
-function getKeyedDataDifference(a, b) {
-  let hasDifference = false;
-  const difference = Object.keys(a).reduce((_difference, key) => {
-    if (!(key in b)) {
-      hasDifference = true;
-      _difference[key] = true;
-    }
-    return _difference;
-  }, {});
-  return hasDifference && difference;
-}
-
-/**
- * Calculate which data-points exist in oldData and not nextData -
- * these are the `exiting` data-points.  Also calculate which
- * data-points exist in nextData and not oldData - thses are the
- * `entering` data-points.
- *
- * @param  {Array} oldData   this.props.data Array
- * @param  {Array} nextData  this.props.data Array
- *
- * @return {Object}          Object with `entering` and `exiting` properties.
- *                           entering[datum.key] will be true if the data is
- *                           entering, and similarly for `exiting`.
- */
-function getNodeTransitions(oldData, nextData) {
-  const oldDataKeyed = oldData && getKeyedData(oldData);
-  const nextDataKeyed = nextData && getKeyedData(nextData);
-
-  return {
-    entering: oldDataKeyed && getKeyedDataDifference(nextDataKeyed, oldDataKeyed),
-    exiting: nextDataKeyed && getKeyedDataDifference(oldDataKeyed, nextDataKeyed)
-  };
-}
-
-function getChildData(child) {
-  if (child.type && child.type.getData) {
-    return child.type.getData(child.props);
-  }
-  return child.props && child.props.data || false;
-}
+import {
+  getDatumKey,
+  getNodeTransitions,
+  getChildData,
+  getChildTransitionDuration
+} from "./transition-helpers";
 
 /**
  * If a parent component has animation enabled, calculate the transitions
@@ -224,13 +176,6 @@ export function getTransitionPropsFactory(props, state, setState) {
       getChildPropsBeforeEnter(animate, data, nodes, () => {
         setState({ nodesShouldEnter: true });
       });
-  };
-
-  const getChildTransitionDuration = function (child, type) {
-    const animate = child.props.animate;
-    const defaultTransitions = child.type && child.type.defaultTransitions;
-    return animate[type] && animate[type].duration ||
-      defaultTransitions[type] && defaultTransitions[type].duration;
   };
 
   return function getTransitionProps(child, index) { // eslint-disable-line max-statements
